requirements.txt
pyarrow==16.0.0 # 16.1.0 segfaults
datasets
openai
tiktoken
anthropic
tenacity
boto3>=1.34.136
jedi
unidiff
python-dotenv
git+https://github.com/castorini/pyserini.git
chardet
GitPython
faiss-cpu
torch
mkl
altair
ipywidgets
beautifulsoup4
googlesearch-python
duckduckgo-search
semantic-text-splitter
sqlite-vss